backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
azure-storage-blob==12.22.0
azure-keyvault-secrets==4.8.0
azure-identity==1.17.1
azure-ai-formrecognizer==3.3.3
azure-cosmos==4.7.0
openai==1.51.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
httpx==0.27.2
applicationinsights==0.11.10
opencensus-ext-azure==1.1.13
aiohttp==3.10.5